+++ Surface_mesh/include/CGAL/draw_surface_mesh.h
@@ -112,11 +112,10 @@ struct Graphics_scene_options_surface_mesh
 
   Graphics_scene_options_surface_mesh(const SM& amesh)
   {
-    bool found=false;
-    std::tie(vcolors, found)=
-        amesh.template property_map<vertex_descriptor, CGAL::IO::Color>("v:color");
-    if(found)
+    auto _vcolors = amesh.template property_map<vertex_descriptor, CGAL::IO::Color>("v:color");
+    if(_vcolors.has_value())
     {
+      vcolors = _vcolors.value();
       this->colored_vertex=[](const SM &, vertex_descriptor)->bool { return true; };
       this->vertex_color=[this](const SM &, vertex_descriptor v)->CGAL::IO::Color
       { return get(vcolors, v); };
@@ -124,10 +123,10 @@ struct Graphics_scene_options_surface_mesh
     else
     { this->colored_vertex=[](const SM &, vertex_descriptor)->bool { return false; }; }
 
-    std::tie(ecolors, found)=
-        amesh.template property_map<edge_descriptor, CGAL::IO::Color>("e:color");
-    if(found)
+    auto _ecolors = amesh.template property_map<edge_descriptor, CGAL::IO::Color>("e:color");
+    if(_ecolors.has_value())
     {
+      ecolors = _ecolors.value();
       this->colored_edge=[](const SM &, edge_descriptor)->bool { return true; };
       this->edge_color=[this](const SM &, edge_descriptor e)->CGAL::IO::Color
       { return get(ecolors, e); };
@@ -135,10 +134,10 @@ struct Graphics_scene_options_surface_mesh
     else
     { this->colored_edge=[](const SM &, edge_descriptor)->bool { return false; }; }
 
-    std::tie(fcolors, found)=
-        amesh.template property_map<face_descriptor, CGAL::IO::Color>("f:color");
-    if(found)
+    auto _fcolors = amesh.template property_map<face_descriptor, CGAL::IO::Color>("f:color");
+    if(_fcolors.has_value())
     {
+      fcolors = _fcolors.value();
       this->colored_face=[](const SM &, face_descriptor)->bool { return true; };
       this->face_color=[this](const SM &, face_descriptor f)->CGAL::IO::Color
       { return get(fcolors, f); };